Added options to assign items in a non-random fashion.
Fixed an issue with Lex's Hero Axe event not triggering.

Fixed an issue with the UI for Skills displaying "Wrath" instead of "Vantage"

Fixed an issue where the conversation items were not all updating properly.

Added the ability to assign holy blood to characters that didn't have it previously.

Added the option to retain horses when randomizing classes.