Added a console log that can be opened with CTRL+SHIFT+C in order ot assist in debugging.
Fixed an issue where patch failure would silently fail instead of notifying the user.

Added logic for FE9 to allow Ragnell and the Regal Sword to be used by the class and not just Ike.

Added logic for FE9 to make sure only Ike has a scripted promotion regardless of his class instead of all Rangers.