Fixed some UI layout issues.
Fixed an issue where the growth min/max spinners were not set correctly if the minimum saved was higher than the default maximum.

Fixed issues with Laguz bosses needing to have AIDs set and the skill to keep them in Beast form.

Disabled logging in the diff compiler to dramatically speed up the randomization for FE9.